# Conflicts:
#	names.txt

#	src/main/java/ca/mcgill/ecse223/quoridor/controller/Quoridor223Controller.java

#	src/test/java/ca/mcgill/ecse223/quoridor/features/CucumberStepDefinitions.java

+++ src/main/java/ca/mcgill/ecse223/quoridor/controller/Quoridor223Controller.java
@@ -52,7 +52,7 @@ public class Quoridor223Controller {
 		
 		// create Quoridor game and get users
 		Quoridor quoridor = QuoridorApplication.getQuoridor();
-		Game newGame = new Game(GameStatus.Initializing, MoveMode.WallMove, quoridor);
+		Game newGame = new Game(GameStatus.Initializing, false, MoveMode.WallMove, quoridor);
 
 	}
 
@@ -694,16 +694,19 @@ public class Quoridor223Controller {
 	 * @author Andrew Ta
 	 * @param playerName player wishes to resign the game
 	 * @throws GameNotRunningException
+	 * @throws GameIsFinished 
 	 */
-	public static void resignGame(String playerName) throws GameNotRunningException{
+	public static void resignGame(String playerName) throws GameNotRunningException, GameIsFinished{
 		if(!isRunning()) throw new GameNotRunningException("Game not running");
 		Game curGame = QuoridorApplication.getQuoridor().getCurrentGame();
 		Player curPlayer = getPlayerByName(playerName);
 		if(curPlayer.equals(curGame.getBlackPlayer())) {
-			curGame.setGameStatus(GameStatus.WhiteWon);
+			curGame.setIsFinished(true);
 		}else {
+			curGame.setIsFinished(true);
 			curGame.setGameStatus(GameStatus.BlackWon);
 		}
+		throw new GameIsFinished("Game Finished: "+curGame.getCurrentPosition().getPlayerToMove().getNextPlayer().getUser().getName() + " won. Congratulation!");
 	}
 	
 	/**
@@ -767,16 +770,17 @@ public class Quoridor223Controller {
 	/**
 	 * @author Le-Li Mao
 	 * @throws InvalidOperationException
-	 * @throws GameIsDrawn 
-	 * @throws GameIsFinished 
-	 * @throws GameNotRunningException 
-	 *
 	 */
-	public static void exitReplayMode() throws InvalidOperationException, GameIsDrawn, GameIsFinished, GameNotRunningException {
+	public static void exitReplayMode() throws InvalidOperationException{
 		if (!isReplay())throw new InvalidOperationException("Not in replay mode.");
 		Game curGame = QuoridorApplication.getQuoridor().getCurrentGame();
-		int ind = findPositionIndex();
 		int size = curGame.getPositions().size()-1;
+		if(curGame.getIsFinished()) {
+			curGame.setCurrentPosition(curGame.getPosition(size));
+			throw new InvalidOperationException("Game Finished: Can't Continue Game");
+		}
+		curGame.setGameStatus(GameStatus.Running);
+		int ind = findPositionIndex();
 		for(int i = ind; i<size;i++) {
 			Move move = curGame.getMove(ind);
 			move.delete();
@@ -784,18 +788,10 @@ public class Quoridor223Controller {
 			GamePosition pos = curGame.getPosition(ind+1);
 			pos.delete();
 		}
-		System.out.println(curGame.getMoves().size());
 		QuoridorApplication.CreateNewWhitePawnBehavior().startGame();
 		QuoridorApplication.CreateNewBlackPawnBehavior().startGame();
-		curGame.setGameStatus(Game.GameStatus.Running);
-		if(identifyDraw()) {
-			throw new GameIsDrawn("Game Draw!");
-		}
-		if(identifyWin()) {
-			throw new GameIsFinished(curGame.getCurrentPosition().getPlayerToMove().getUser().getName() + " won");
-		}
+		
 	}
-	
 	public static void jumpToFinalPosition() throws InvalidOperationException {
 		if (!isReplay())
 			throw new InvalidOperationException("Game is not in replay mode");
@@ -807,7 +803,7 @@ public class Quoridor223Controller {
 	}
 	
 	/**
-	 * Perform a step forward in the 
+	 * Perform a step forward in the replay mode
 	 * @author Le-Li Mao
 	 * @throws InvalidOperationException
 	 */
@@ -818,7 +814,11 @@ public class Quoridor223Controller {
 		if(ind==curGame.getPositions().size()-1)throw new InvalidOperationException("Already at the last step");
 		curGame.setCurrentPosition(curGame.getPosition(ind+1));
 	}
-	
+	/**
+	 * Perform a step backward in the replay mode
+	 * @author Le-Li Mao
+	 * @throws InvalidOperationException
+	 */
 	public static void StepBackward() throws InvalidOperationException {
 		if (!isReplay())throw new InvalidOperationException("Game is not in replay mode");
 		Game curGame = QuoridorApplication.getQuoridor().getCurrentGame();
@@ -1317,39 +1317,6 @@ public class Quoridor223Controller {
 		return isWallPositionValid(target_tile.getRow(), target_tile.getColumn());
 	}
 	
-	
-	// @sacha: is the wall side to indicate the direction a good way of making transitions for pawn ?
-	/**
-	 * MovePlayer feature dev 
-	 * @author mixed
-	 * @param side
-	 * 
-	*/
-	/*public static void movePlayer(TOWall.Side side) throws GameNotRunningException, InvalidOperationException {
-		if (!isRunning()) throw new GameNotRunningException("Game not running");
-		Game current_game = QuoridorApplication.getQuoridor().getCurrentGame();
-		Board current_board = QuoridorApplication.getQuoridor().getBoard();
-		
-		// get player moving
-		PlayerPosition current_position;
-		if(isWhitePlayer()) current_position = current_game.getCurrentPosition().getWhitePosition();
-		else current_position = current_game.getCurrentPosition().getBlackPosition();
-
-		int newRow = current_position.getTile().getRow()
-				+ (side == TOWall.Side.Up ? -1 : side == TOWall.Side.Down ? 1 : 0);
-		int newCol = current_position.getTile().getColumn()
-				+ (side == TOWall.Side.Left ? -1 : side == TOWall.Side.Right ? 1 : 0);
-		// first case simple legal position check (does this position exists on board ?) 
-		if (!isWallPositionValid(newRow, newCol)) throw new InvalidOperationException("Illegal Move");
-		if (!is
-veLegal(newRow, newCol)) throw new InvalidOperationException(String.format("%s: Invalid move, try again !", getCurrentPlayerName()));
-		
-		// might need to get the next tile using indexes & get from tiles list in board
-		Tile next_tile = getTile(newRow, newCol);
-		current_position.setTile(next_tile);
-		SwitchPlayer();
-	}*/
-	
 	/////////////////////////////////////////////////////
 	//////////// Move Pawn and Jump Pawn/////////////////
 	/////////////////////////////////////////////////////
@@ -1445,36 +1412,14 @@ veLegal(newRow, newCol)) throw new InvalidOperationException(String.format("%s: 
 				}
 			} 
 		}
-		//Player current_player = current_game.getCurrentPosition().getPlayerToMove();
-		//PlayerPosition current_position;
-		//PlayerPosition opponent_position;
-		
-		//if(current_player.equals(current_game.getBlackPlayer())) {
-		//	current_position = current_game.getCurrentPosition().getBlackPosition();
-		//	opponent_position = current_game.getCurrentPosition().getWhitePosition();
-		//}
-		//else {
-		//	current_position = current_game.getCurrentPosition().getWhitePosition();
-		//	opponent_position = current_game.getCurrentPosition().getBlackPosition();
-		//}
-		
-	/*if(identifyDraw()) {
-			// TODO: may need another switch player here for when loading a draw game
-			throw new GameIsDrawn("Game Draw!");
-		}else if(identifyWin()) {
-		String winText = curGame.getCurrentPosition().getPlayerToMove().getUser().getName() + " won. Congratulations!";
-		SwitchPlayer();
-		throw new GameIsFinished(winText);
-	}else {
-		SwitchPlayer();
-	*/
-		
+
 		if(identifyDraw()) {
-			// TODO: may need another switch player here for when loading a draw game
-			throw new GameIsDrawn("Game Draw!");
+			curGame.setIsFinished(true);
+			throw new GameIsDrawn("Game Finished: Game Draw!");
 		}
 		if(identifyWin()) {
-			throw new GameIsFinished(curGame.getCurrentPosition().getPlayerToMove().getUser().getName() + " won. Congratulation!");
+			curGame.setIsFinished(true);
+			throw new GameIsFinished("Game Finished: "+ curGame.getCurrentPosition().getPlayerToMove().getUser().getName() + " won. Congratulations!");
 		}
 		SwitchPlayer();
 	}
